requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
pytest
azure-identity
azure-keyvault-secrets
azure-keyvault-keys